requirements.txt
fastapi
uvicorn
python-multipart
librosa
soundfile
torch
tensorflow
numpy
pillow
transformers
datasets
openai
python-dotenv
scipy
